Better way to extract price and amount in tickets. Now amount is a double value

+++ src/main/java/me/renedo/payment/app/http/receipt/record/LineResponse.java
@@ -6,7 +6,7 @@ import java.util.UUID;
 import me.renedo.payment.line.domain.Line;
 import me.renedo.shared.date.ISOFormatter;
 
-public record LineResponse(UUID id, UUID item, String name, BigDecimal total, Integer amount, String updated){
+public record LineResponse(UUID id, UUID item, String name, BigDecimal total, Double amount, String updated){
     public LineResponse(Line line){
         this(line.getId(), line.getItem(), line.getName(), line.getTotal(), line.getAmount(),  ISOFormatter.format(line.getCreated()));
     }
